Fix URI handling for WireserverTransport
The previous handling adhered to the request mutation guidelines
outlined in "NMAgent - Swift Provisioning". However, these do not track
with what works in practice. A working implementation URL encodes the
"type" portion of the query string, which makes sense. It's possible
that the documentation relied on more forgiving behavior from cURL in
the examples provided.

This introduces a new type to handle the construction of the query
string so that rules can be applied consistently.