Ensure required parameters are set for PublishNC
The PutNetworkContainerRequest type depended on several fields that
were not enforced by the Validate method. If these fields aren't
provided, it produces an invalid URL that results in a confusing and
obtuse error from NMAgent. The problem can be checked at the client
level and produce a more helpful error. This is accomplished here.

Also, updated the usage of PutNetworkContainerRequest to adhere to these
requirements. Furthermore, there were several JSON struct tags that are
no longer necessary because there is an intermediary type that handles
encoding/decoding to/from JSON.